[Pal] Move header files under include/ and adjust the include directive
