+++ Pal/include/host/Linux-SGX/host_endian.h
