+++ Pal/include/host/Skeleton/host_endian.h
