+++ Pal/src/host/Skeleton/Makefile
@@ -1,7 +1,8 @@
 include ../../../../Scripts/Makefile.configs
 include Makefile.am
 
-CFLAGS  += -I. -I../.. -I../../../include -I../../../lib
+CFLAGS  += -I. -I../.. -I../../../include  -I../../../include/pal \
+	-I../../../include/host/Skeleton -I../../../lib
 ASFLAGS += -I. -I../.. -I../../../include
 
 host_files = libpal-Skeleton.a pal.map
